mpt.c: mpt.h: 	Add support for reading extended configuration pages. mpt_cam.c: 	Do a top level topology scan on the SAS controller.  If any SATA 	device are discovered in this scan, send a passthrough FIS to set 	the write cache.  This is controllable through the following 	tunable at boot:
	hw.mpt.enable_sata_wc:
		-1 = Do not configure, use the controller default
		 0 = Disable the write cache
		 1 = Enable the write cache

	The default is -1.  This tunable is just a hack and may be
	deprecated in the future.

Turning on the write cache alleviates the write performance problems with
SATA that many people have observed.  It is not recommend for those who
value data reliability!  I cannot stress this strongly enough.  However,
it is useful in certain circumstances, and it brings the performence in line
with what a generic SATA controller running under the FreeBSD ATA driver
provides (and the ATA driver has had the WC enabled by default for years).


git-svn-id: svn://svn.freebsd.org/base/head@170252 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f